edi: Allow for timeouts to be specified for an EDI gateway
Allow for an EDI gateway to specify a timeout value, and use this for
the SSH connection and banner timeouts, and for the SFTP command
channel timeout.  This allows stuck connections to be terminated
without relying on the ir.cron overall job timeout (which is likely to
be set to a high value to allow for long-running background jobs).

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>