[940] Migrate EDI module to Odoo11 + some fixes
Do not store the password in the database, load it from the config file
instead.
- Renaming fields to remove "x_".
- Using odoo-core-provided config functions.
- Adding [edi] section into the config file that stores the passwords.

Optional deferred execution of documents.
- Renaming fields to remove "x_".
- Incorporating the deferred logic directly into the do_transfer function,
instead of changing the behaviour of document preparation.
- If an error occurs in deferred execution of a document (e.g. trying to
reserve a split package) and during the deferred execution the gateway
has been changed (e.g. transferred another file), a serilisation error
occurs when the error from deferred execution is recorded, hiding the
initial error. To avoid this when errors occur during deferred execution
they are not logged on the gateway, so serialisation errors do not occur.

Adding field to edi.gateway to record if gateway is for production use.
- Gateways only transfer if the value of their is_production_gateway
field matches that of the 'production' flag in the [edi] section of the
config file.