edi: Use sudo() when posting on associated threads
The user raising an issue for an EDI document may not have permission
to modify the associated EDI transfer or EDI gateway.  Similarly, the
use raising an issue for an EDI transfer may not have permission to
modify the associated EDI gateway.

Fix by using sudo() when posting the messages to the threads for the
associated EDI transfer and/or EDI gateway.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>