edi: Allow gateway password to be stored in Odoo configuration file
Storing the password in the database is potentially undesirable, not
least since a production backup that is restored onto a development
machine for debugging will include the production gateway password(s).

Fix by allowing the gateway password to be retrieved from the Odoo
configuration file, with the gateway record merely defining the name
of the configuration file option from which the password should be
retrieved.

Originally-implemented-by: Charlie Wheeler-Robinson <crwr@pigotts.org.uk>
Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>