edi: Expose document's record type names to view rendering logic
When rendering an EDI document, the "records" notebook should contain
a page for each applicable record type.  Determining the applicable
record types is non-trivial since the domain syntax does not allow us
to express the concept of "rec_type_ids contains <x>".

One possible approach is for each record type to define a boolean
"use_<record_type>" flag as an extension to edi.document.type, and
make this flag available in edi.document as a non-stored related
field.  This involves a substantial amount of boilerplate code for
each record type, which is undesirable.

Fix by constructing a slash-separated list of record model names as a
string field on edi.document.type (computed from rec_type_ids), and
expose this field as a non-stored related field in edi.document.  This
allows record types to define a notebook page using code such as:

  <xpath expr="//notebook[@name='records']" position="inside">
    <page name="special_record_type" string="Special records"
          attrs="{'invisible':['!',('rec_type_names','like',
                                    '/edi.record.special/')]}">
      ...
    </page>
  </xpath>

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>