edi: Add batched() helper function
A common requirement when processing EDI documents is to process
records in batches, to obtain a sensible trade-off between the number
of database queries and the size of each query.

Add a batched() helper function which takes an iterable and constructs
subsequences of a specified size, along with a range object to allow
for easy progress logging by the caller.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>