[IMP] edi: Inhibit EDI issue exception logging during testing
When the EDI code catches an exception and converts it to an EDI
issue, the exception is also logged in the normal way to provide
visibility of the exception in the system log files.

Automated tests are likely to deliberately provoke exceptions.  The
explicit logging performed as part of EDI issue creation results in
unwanted and potentially distracting or misleading noise within the
automated test log.

Fix by inhibiting the explicit logging while automated testing is
in progress.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>