[TEST] edi: Report unexpected raised issues as cause of assertion failures
When assertRaisesIssue() fails, it is useful to be able to see the
original exception (which has already been converted to an EDI issue)
in order to determine the cause of the test failure.

Capture the exception(s) passed to edi.issues.raise_issue() within the
scope of assertRaisesIssue() and attach the original exception as the
root cause of any test failure.  If multiple issues were raised, we
attempt to guess which is most likely to be the root cause.

This allows us to remove the code that currently unconditionally hides
the output of _logger.error() during unit test runs.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>