[TEST] edi: Assert that "added host key" warning is generated
Check that the "Added host key" warning is generated when using
ssh_connect() with a previously unknown host key.

As a pleasant side effect, this removes the mildly distracting warning
from the test log.

Note that we cannot use assertWarns() since calling _logger.warning()
directly does not count as a warning.  (We cannot viably use warn() to
generate the warning message since there seems to be no sensible way
to produce a warning without a filename and line number.)

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>