[TEST] edi: Fix ability to retest "edi" after installing dependent modules
If a dependent module such as "edi_product" has already been
installed, then an attempt to reinstall and test the "edi" module will
fail with a KeyError in edi.document.type.autocreate() for the line:

   Model = self.env[doc_type.model_id.model]

This appears to be because the model classes for the dependent modules
are not created until the tests for the "edi" module have completed.

Resolve this unexpected situation by deleting any document types that
correspond to non-existent model classes.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>