[IMP] edi_product: Prepare for generalisation of edi.product.document
Prepare for the imminent generalisation of edi.product.document to a
class capable of synchronizing records in an arbitrary Odoo model, by
adding the substring "product" to the "_record_values" method.

This will be required in order to support inheritance from multiple
document models, each inheriting from the base synchronizer document
model.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>