[IMP] edi: Add generic concept of an EDI lookup relationship
A common pattern within EDI document processing is for EDI records to
refer to other database records using a lookup key.  For example: a
stock picking request from an external system may refer to products
using the product code.

This will be represented within the EDI record as a pair of fields:
one holding the lookup key value (e.g. the product's "default_code")
and the other holding the target record itself (e.g. the
"product_id").

The actual target record may be created during the execution of the
EDI document; the target record field may therefore be empty when the
EDI record is created.

The "edi_relates" field attribute can be used to annotate a field as a
lookup key, using a syntax similar to the standard "related" field
attribute.  For example, if a model defines the fields "product_key"
and "product_id", then the "product_key" field definition would
include the attribute

  edi_relates='product_id.default_code'

to indicate that the EDI core should populate the "product_id" column
based on searching for "product.product.default_code" equal to the
value of the "product_key" column.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>