[IMP] edi: Extend BaseModel to include sliced() and batched() methods
EDI code typically deals with large number of records, which are often
most efficiently processed in batches of PREFETCH_MAX records at a
time in order to minimise the number of database round trips.

The batched() iterator helper can break a recordset down into batches
of the appropriate size, but the returned batch object ends up being a
list of singleton recordsets rather than a standard recordset.  It is
therefore not possible to call recordset methods directly on the
returned batch object.

Extend the Odoo base model class to add sliced() and batched()
methods, using an approach similar to the Odoo base_sparse_fields
addon.  These new methods are conceptually equivalent to the sliced()
and batched() iterator helpers, but return recordset objects instead
of lists.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>