[FIX] edi: Behave correctly when called with a non-empty recordset
The @api.model decorator merely indicates that the method should
ignore the contents of the recordset; it does not actually cause the
method to be called with an empty recordset.

Various code paths currently assume that an @api.model method will
always be called with an empty recordset and use self[fieldname] to
obtain an empty recordset for the target of a relational field.  This
will fail if the original recordset is non-empty.

Fix by using self.browse()[fieldname] instead of self[fieldname].

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>