[IMP] edi_stock: Extend move/pick requests to handle updates
Move requests have been extended to be able to update or cancel stock
moves in addition of creating them. Pick requests have been updated to
reuse exisiting pickings. The criteria to find the moves and picks is by
matching the request name with the move name and the pick origin
respectively.

Move requests to update and cancel are not allowed for started pickings,
i.e., pickings with move lines with qty_done > 0. This does not include
pickings in state done or cancel.

Signed-off-by: Miquel PS <miquel.palahi.sitges@unipart.io>