[IMP] edi_stock: Change edi.pick.request.record to use a synchronizer
Use a synchronizer approach to handling edi.pick.request.record.  This
allows us to remove the explicit foreign key relationship between
edi.move.request.record and edi.pick.request.record (which causes
additional work when subclassing either of these classes), and allows
for move request records to refer to existing stock.picking records
without requiring the existence of a corresponding pick request
record.

Default to filtering out cancelled stock.picking records, on the basis
that an EDI document for an already-cancelled picking probably
represents a situation in which the operation has decided to cancel
the picking due to a fault in the original EDI document, and is now
attempting to recreate the picking after fixing the problem with the
EDI document.  Subclasses (such as the pick request tutorial) may
choose to modify this filtering policy.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>