[IMP] edi_stock: Clean up modification of existing stock transfers
Implement an alternative design for the modification of existing stock
transfers, intended to allow for more flexible policy choices by
subclasses.

The edi.move.request.record.existing_move() method is used to identify
the corresponding existing move, if any.  The default action is to
identify uncancelled moves with the same tracking identity, product,
and picking type.  Subclasses may override this behaviour by
overriding the existing_move() method and applying extra filtering
and/or validation rules.

The concept of "action" is removed from the edi.move.request.record
class.  The distinction between "create" and "update" is determined by
whether or not the existing_move() method returns an existing move,
and a quantity of zero is used to indicate "delete".  Subclasses may
override this behaviour as needed.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>