[DOC] edi: Note that record model detection methods need not be overridden
The purpose of the xxxx_record_model() methods is to automatically
determine the applicable record model subclass based on the document
type's associated record types.  There is generally no need for
individual document models to override these methods.

Make this explicit within the documentation, since it is a common
mistake for developers to assume that these methods must be overridden
in order to pick up the correct record class.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>