[IMP] edi: Always populate EDI lookup relationship target record field
Always fill in the target field for an EDI lookup relationship within
a list of value dictionaries.  This removes the requirement for
individual EDI record models to handle the possibility that a lookup
relationship may fail to find a target record at the point of document
preparation (e.g. because the target record will be created by another
EDI record class for the same document).

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>