[FIX] edi: Avoid deactivating all target records when helpers are not used
EDI synchronizer document models tend to provide a convenience method
for constructing an iterable of EDI record field value dictionaries,
and a prepare() method that calls this convenience method and passes
the results to the record model's prepare() method.  This typically
eliminates the need for derived synchronizer document models to
provide a boilerplate prepare() method; the derived model may instead
simply implement the convenience method.

This causes problems when the record model has an associated
deactivator.  The convenience method will always be called even if the
derived document model has chosen not to use it, with the end result
being that every possible target record will be deactivated.

Fix by allowing the base document model's placeholder convenience
method to raise a special-purpose exception as a means of signalling
to the record model's prepare() method that it is an unused method
(rather than being a method that just happens to be returning an empty
iterable).

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>