[IMP] edi: Allow raw document importer to handle read-only fields
The previewer in base_import will never match against read-only
fields, resulting in an import failure for models such as stock.quant
where almost all useful fields are read-only.

Fix by using a trial import (with a rollback to a savepoint) as the
document preparation stage, instead of using the previewer's match
list to determine validity.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>