[TEST] edi: Do not assume success of raw document preparation with bad data
Raw documents are essentially validated only at the point of
execution; the minimal validation performed at preparation time exists
primarily to verify that the header list is recognised.  Since this
preparation-time validation uses an essentially undefined subset of
the imported data, no assumptions should be made as to whether or not
the preparation will fail due to an invalid data value.

Fix by asserting only that the document with invalid data will fail at
the point of execution.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>