[FIX] edi: Defer document creation until autodetection order is known
Sorting the autocreated document recordset is insufficient since this
ordering will be lost when the recordset is stored in a Many2many
field.

Fix by separating document creation from document type detection, so
that the documents will be created in the intended order and hence
maintained in this order upon subsequent retrieval from the database.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>