[IMP] edi: Allow groupby() to use a recordset as the sorting key
Using groupby() with a key that produces a recordset (e.g. a Many2one
field name) will currently result in undefined behaviour since
singleton recordset objects cannot be compared for ordered inequality
using the standard comparison operators.

The standard point-of-use workaround is to use a key that produces a
recordset ID instead of the recordset object itself.  This results in
cumbersome code that must perform an explicit browse() to reconstruct
the desired singleton recordset.

Support the use of a recordset as the sorting key by checking for a
key that produces a recordset object and modifying the sort key to
automatically use the (assumed singleton) recordset's ID.  The
grouping key is left unmodified, since singleton recordsets may safely
be compared for equality using the standard comparison operators.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>