[IMP] edi_stock, edi_sale: Allow reporting document field to be overridden
The stock.picking.edi_pick_report_id and sale.order.edi_sale_report_id
fields are used to record the EDI documents used to report upon the
stock.picking and sale.order objects respectively, on the assumption
that such objects need to be reported upon only once.

Allow these reporting EDI document fields to be overridden by the EDI
document model, in order to accommodate situations in which it may be
necessary for multiple document models to independently report upon
the same objects.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>