[IMP] edi_sale, edi_stock: Use superclass record model prepare() method
The various reporting record types tend to use a prepare() method that
accepts an iterable of recordsets to be reported upon.  This differs
from the superclass prepare() method, which accepts an iterable of
value dictionaries for records to be created.

Improve this situation by calling the superclass prepare() method,
translating the iterable of recordsets to an iterable of record value
dictionaries via the record_values() method.  This is not perfectly
clean (since the nature of the elements within the iterable differs
between subclass and superclass), but is an improvement upon the
current situation in which the subclass method is entirely unrelated
to the superclass method.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>