[IMP] edi: Expose document preparation date during call to prepare()
Some document models may wish to use the document preparation date as
part of the preparation process: for example, to filter records based
on the preparation date.  Allow for this by setting the preparation
date before calling prepare().  Do so within the savepoint so that the
preparation date is automatically rolled back if an exception occurs.

This modifies the semantics of the preparation date field to reflect
the time that the process was started rather than the time that the
process completed.  This therefore represents the timestamp of the
database snapshot used when preparing the document, which is probably
more generally useful than the current semantics.

The execution date field is left to represent the time that the
execution process completed.  This seems more generally useful since
it represents the time that changes made by the document were
committed to the database.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>