[IMP] edi: Automatically elide duplicate EDI synchronizer records
The existing synchronizer target comparison check will elide any EDI
records that would not result in a new or modified target Odoo record.
If the document model attempts to create two identical EDI records,
neither of which would result in a new or modified target Odoo record,
then both EDI records will be elided.

However, if the document model produces the same two identical EDI
records in a situation where the target Odoo record does not yet
exist, then neither EDI record will be elided.  This will result in a
database constraint violation, since the synchronizer requires that
the EDI record name is unique within a document.

Make the behaviour less surprising by automatically eliding any
duplicate EDI records that are not already elided by the target
comparison check.  This avoids the constraint violation in the second
of the above scenarios.

There is a performance and memory cost to maintaining the lookup cache
required to perform this deduplication.  This cost is minimised by
placing the deduplication check after the target comparison check.  In
a typical use case for a synchronizer document, all EDI records will
be produced only once and the vast majority of EDI records will be
elided by the target comparison check.  By performing the target
comparison check first, the deduplication check needs to operate on
only the small fraction of EDI records that are not already elided.

Where the document model guarantees not to produce duplicate EDI
records, the EDI record model may reduce the performance and memory
cost further by setting the _edi_sync_dedupe flag to False: this will
eliminate the deduplication check entirely.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>