[IMP] edi: Detect missing or inadequate precaching for synchronizer models
Synchronizer models should provide adequate precaching to ensure that
the number of queries is substantially fewer than one per record.
Detect when this condition is not met and display a warning message.

Most derived models will not need to perform any explicit precaching,
since the EDI lookup relationship mechanism and the EDI synchronizer
itself both take care to minimise the number of queries.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>