[FIX] edi: Set edi.document.project_id to read only
The field is a related field to doc_type_id.project_id, which
is triggering a write to doc_type_id every time a new document
of that type is created.
This causes issues for users that only have read access rights
to edi.document.type model.

Setting the field as readonly prevents the write to doc_type_id.

Signed-off-by: Miquel PS <miquel.palahi.sitges@unipart.io>