[IMP] edi: Clarify meaning of synchronizer query count debug message
The query count reported by edi.record.sync.elide() is actually the
number of queries in excess of the number of non-elided records (on
the basis that each created EDI record will require at least one
INSERT query).

Clarify this meaning in the choice of variable names and in the debug
message text.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>