[IMP] edi: Defer computation during document preparation and execution
Add "recompute=False" to the context used during EDI document
preparation and execution.  This can significantly reduce the number
of database updates required: a typical reduction is from once per
computed field per record down to once per record.

This cost of this reduction is the memory occupied for the list of
pending recomputations.  Keep this memory usage under control with
explicit calls to self.recompute() at appropriate junctures.

A good heuristic is to include a call to self.recompute() at the end
of any code block that uses self.statistics(), since the use of
self.statistics() would typically indicate a block of code that is
expected to perform a reasonable chunk of work.

Despite this heuristic, we do not explicitly call self.recompute()
within self.statistics(), since doing so would introduce side effects
into what appears to the reader to be purely diagnostic code.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>