[IMP] edi-sale: add capability to cancel requests
Here we add a new method to the EdiSaleRequestDocument
interface that will enable filtering request orders
that must be cancelled (insted of confirmed) in case
the _auto_confirm attr is flagged.

Such capability is necessary for those subclasses
that need to differentiate between requests to be
confirmed and cancelled immediately, while keeping
the usual processing based on batches of records.

User-story: 2811