[IMP] edi: Provide all synchronizer target record default values
Odoo's methods for calculating default values tend to be exorbitantly
expensive: it is not unusual to find default value methods that will
perform an identical database search for each new record.

Work around this persistent flaw by constructing the default values
dictionary at most once, and then merging the default values in to the
values dictionaries passed to create().

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>