[IMP] edi, edi_sale, edi_stock: Formalise precaching mechanism
The prefetch mechanism is not sufficient to guarantee that all queries
are performed in bulk.  For example, consider the code:

    for rec in batch:
        print(rec.order_id.partner_id.name)

Since this code uses only iteration and field traversal by attribute
access, the same prefetch object will be used throughout.

On the first iteration, evaluating "rec.order_id" should therefore
fetch the field "order_id" (along with all other prefetchable fields)
for all records in "batch", and add all of the "order_id" record IDs
to the prefetch object.

Similarly, evaluating ".partner_id" should fetch the field
"partner_id" (along with all other prefetchable fields) for all
applicable "order_id" records, and add all of the "partner_id" record
IDs to the prefetch object.

Finally, evaluating ".name" should fetch the field "name" (along with
all other prefetchable fields) for all applicable "partner_id"
records.

All fetches are therefore performed as bulk operations on the first
iteration of the loop, and all subsequent loop iterations therefore
experience a 100% cache hit rate.

A problem occurs if some of the intermediate "order_id" or
"partner_id" values are already present in the cache.  On the first
iteration, if "rec.order_id" is already present in the cache then
evaluating "rec.order_id" will just return the cached value, and will
add only this "order_id" record ID to the prefetch object.

Subsequently evaluating ".partner_id" will fetch fields for only the
single "order_id" record, since that is the only record present in the
prefetch object.

The net result is that code expected to exploit the prefetch mechanism
to perform bulk fetch operations may end up degenerating to perform
multiple single-record fetch operations.  This degeneration is
unfortunately dependent on the a priori cache state and so cannot
viably be predicted.

The prefetch mechanism must therefore be seen as a heuristic that will
attempt to perform bulk operations but cannot guarantee to do so, even
with careful management of the prefetch object.

Revert therefore to using explicit precaching, with sensible defaults
and well-defined injection points for derived models.

Note that the ".product_tmpl_id.name" is not strictly necessary, since
the same effect could be achieved via just ".name".  The explicit
"product_tmpl_id" is included to make it easier for the reader to
understand why this extra precache fetch is required.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>