[IMP] edi: Current document activity status available
Adding an activity status to edi.document to contain the
current activity (preparing, executing) of the document.

EDI gateway transfers commit explicitly so that each
stage of transfer>prepare>execute are visible, a
rollback will only occur for a subsection of this
transcation - a failed execution will now  leave a
prepared document that can be viewed in the UI.

Using clear_upon_failure() to prevent recomputes
after an exception.

Co-authored-by: Kevin Dwyer <kevin.dwyer@unipart.io>