[FIX] edi: Add the concept of target not yet known for edi relates field
A synchronizer will not recognise a change in an edi related field, if that
field is not set beforehand. By marking the field as "target not yet
known" this signals to the synchronizer that this value will be a new record
after execution and therefore should not be elided.

User-story: story/4349
Signed-off-by: Michele Costa <michele.costa@unipart.io>