[IMP] edi: Restrict local filesystem gateway to specific path
If a user had permission to modify the paths of the local filesystem
gateway they were able to retrieve files from the server, provided
the user running the server has access to them.

This change adds a restriction to local filesystem connections that
they can only access files within a pre-defined path. Changing that
path requires modifying the config file and restarting the service.

User-story: story/4114