[FIX] edi: Fix placement of assertion in tearDown() method
The EDI gateway base test case asserts that no issues were raised as
part of tearDown(), but does so before calling the superclass
tearDown() method.  If this assertion fails, the superclass method
will not be called and so the transaction will never be rolled back,
resulting in the failure of all remaining tests in the suite.

Fix by delaying the assertion until after calling the superclass.

Story: 4392
Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>