[FIX] edi_sale: Avoid unintentional dependency on "stock" module
The EDI sale base test case defines three test products with a type of
"product" (i.e. stockable product, rather than consumable).  This type
is defined only by the "stock" module, which is deliberately not a
dependency of the "edi_sale" module.

The EDI sale test cases do not require the products to be stockable,
so fix by simply omitting the product type when creating the product.

This problem is not detected by the standard CI test run since the
standard test setup installs all EDI modules including "edi_stock",
which drags in the "stock" module and so hides the problem.

Reported-by: Luke Barone-Adesi <luke.barone-adesi@unipart.io>
Reported-by: Przemysław Buczkowski <pbuczkowski@unipart.io>
Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>