[ADD] edi: Fix issue with elide relating to NewId
Recent changes to fix issues with comparators has introduced a
new edge case. Two identical values may be evaluated as different
when part of them (while being the same value) needs to be
constructed as another new record. Example: partner title record
has not yet been constructed and the same partner is attempted to
be created twice, this results in the partner being identified as
different causing an attempt at creating the parnter twice, the
unique constraint for a partner then causes the creation of a
partner to be invalidated.

Signed-off-by: Robert Smith <robert.smith@unipart.io>