[IMP] edi_sale: Skip invalid orders and order lines
User-story/10956
Task/11085

Signed-off-by: Kevin Dwyer <kevin.dwyer@unipart.io>

If we are not in fail-fast mode, do not process invalid order lines or
lines with no order (because we could not create the order).

Delete orders which are left with no lines.

Delete partners with no orders, as long as they do not have any historic
orders.