[IMP] edi: Add fail_fast and error handling
Add fail fast option for EDI document types and error handling in
for EDI records. This commit does not add the logic which handles
cases where the fail fast is disabled for each document type.
This needs to be implemented in each document type model.

The logic for fail fast and error field for edi records needs to
be implemented with each document type that handles fail fast
disabled. Sale request document tutorial will be one of the first
document types to implement this logic and add tests as well (in
upcoming commits).

User-story: x2424

Signed-off-by: Lorenzo Cucurachi <lorenzo.cucurachi@unipart.io>