[IMP] edi_sale: Sale Order, add fail_fast logic, tutorials and tests
Add the logic that handles fail fast option disabled, tutorial documents
which are needed for tests and tests to check the logic works as expected.

The sale order tutorial document type and record types have been
marked as active=False in the data files in order to avoid displaying
these to the user. Theese are activated in the tests in order to allow
the tests to run the edi code on an example document.

User-story: x2424

Signed-off-by: Lorenzo Cucurachi <lorenzo.cucurachi@unipart.io>